+++ src/LaunchMsaView/components/NewNCBIBlastQuery/types.ts
@@ -1,0 +1,11 @@
+export interface Feat {
+  start: number
+  end: number
+  type?: string
+}
+
+export interface SeqState {
+  seq: string
+  upstream?: string
+  downstream?: string
+}
